In the process . . .

+++ include/ft_printf.h
@@ -20,7 +20,7 @@ int		ft_printf(const char *format, ...);
 int		ft_formats(va_list args, const char format);
 int		ft_printchar(int c);
 int		ft_printstr(char *str);
-int		ft_print_ptr(unsigned long long ptr);
+int		ft_print_ptr(unsigned long ptr);
 int		ft_printnbr(int n);
 int		ft_print_unsigned(unsigned int n);
 int		ft_print_hex(unsigned int num, const char format);
